+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessFilesTask.java
@@ -20,6 +20,7 @@ package com.samaxes.maven.minify.plugin;
 
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -104,11 +105,12 @@ public abstract class ProcessFilesTask implements Callable<Object> {
      * @param outputFilename the output file name
      * @param engine minify processor engine selected
      * @param yuiConfig YUI Compressor configuration
+     * @throws FileNotFoundException when the given source file does not exist
      */
     public ProcessFilesTask(Log log, boolean verbose, Integer bufferSize, String charset, String suffix,
             boolean nosuffix, boolean skipMerge, boolean skipMinify, String webappSourceDir, String webappTargetDir,
             String inputDir, List<String> sourceFiles, List<String> sourceIncludes, List<String> sourceExcludes,
-            String outputDir, String outputFilename, Engine engine, YuiConfig yuiConfig) {
+            String outputDir, String outputFilename, Engine engine, YuiConfig yuiConfig) throws FileNotFoundException {
         this.log = log;
         this.verbose = verbose;
         this.bufferSize = bufferSize;
@@ -250,8 +252,9 @@ public abstract class ProcessFilesTask implements Callable<Object> {
      *
      * @param finalFilename the final file name
      * @param sourceFilename the source file name
+     * @throws FileNotFoundException when the given source file does not exist
      */
-    private void addNewSourceFile(String finalFilename, String sourceFilename) {
+    private void addNewSourceFile(String finalFilename, String sourceFilename) throws FileNotFoundException {
         File sourceFile = new File(sourceDir, sourceFilename);
 
         addNewSourceFile(finalFilename, sourceFile);
@@ -262,8 +265,9 @@ public abstract class ProcessFilesTask implements Callable<Object> {
      *
      * @param finalFilename the final file name
      * @param sourceFile the source file
+     * @throws FileNotFoundException when the given source file does not exist
      */
-    private void addNewSourceFile(String finalFilename, File sourceFile) {
+    private void addNewSourceFile(String finalFilename, File sourceFile) throws FileNotFoundException {
         if (sourceFile.exists()) {
             if (finalFilename.equalsIgnoreCase(sourceFile.getName())) {
                 log.warn("The source file [" + ((verbose) ? sourceFile.getPath() : sourceFile.getName())
@@ -272,8 +276,8 @@ public abstract class ProcessFilesTask implements Callable<Object> {
             log.debug("Adding source file [" + ((verbose) ? sourceFile.getPath() : sourceFile.getName()) + "].");
             files.add(sourceFile);
         } else {
-            log.warn("The source file [" + ((verbose) ? sourceFile.getPath() : sourceFile.getName())
-                    + "] does not exist.");
+            throw new FileNotFoundException("The source file ["
+                    + ((verbose) ? sourceFile.getPath() : sourceFile.getName()) + "] does not exist.");
         }
     }